Scoring: Fix error in engagement score calculation
Engagement Points (points for the Yellow Section) are NOT allowed to be
included in the engagement bonus calculation.  The mission text specifies
that it applies only to, "Non-Engagement Mission Score Total".  So you
need to be sure to remove points for the "Yellow Section" part of
Engagement prior to calculating the bonus, and then add the points back
after applying the bonus.